backend/requirements.txt
# Core
python-telegram-bot==20.7
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Utils
python-dotenv==1.0.0
httpx==0.26.0
python-multipart==0.0.6

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==23.12.1
mypy==1.8.0
